Updated API to remove hardcopy of file from directory ./uploads

+++ src/uploads/uploads.service.ts
@@ -1,7 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { Uploads } from './entities/uploads.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Like, Repository } from 'typeorm';
+import * as fs from 'fs';
+import * as path from 'path';
 
 @Injectable()
 export class UploadsService {
@@ -34,9 +36,31 @@ export class UploadsService {
     }
 
     async deleteFile(fileId: number) {
+        const file = await this.uploadsRepository.findOneBy({id: fileId});
+        if (!file) {
+            throw new BadRequestException('Invalid file id');
+        }
+        // first delete file from folder
+        this.deleteFileFromDirectory(file.fileName);
+        // second delete file info from DB table
         const result = await this.uploadsRepository.delete(fileId);
         return {
             message: result.affected ? 'Deleted successfully' : 'File with id:'+fileId+' not found'
         }   
+    }
+
+    private deleteFileFromDirectory(fileName: string): string {
+        const filePath = path.join(__dirname, '..', '..', 'uploads', fileName); // Adjust path as needed
+
+        if (!fs.existsSync(filePath)) {
+            return `File "${fileName}" not found.`;
+        }
+
+        try {
+            fs.unlinkSync(filePath);
+            return `File "${fileName}" deleted successfully.`;
+        } catch (err) {
+            throw new Error(`Error deleting file: ${err.message}`);
+        }
     }
 }
